Change BluetoothGattService::GattErrorCode to an enum class.
Change BluetoothGattService::GattErrorCode to enum classes to avoid
conflicts in enum values, as detected by -Wshadow.

Bug: 1344231
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3906106
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1049227}

+++ ash/services/secure_channel/ble_characteristics_finder.cc
@@ -184,7 +184,7 @@ void BluetoothLowEnergyCharacteristicsFinder::OnRemoteCharacteristicRead(
 
   if (error_code.has_value()) {
     PA_LOG(ERROR) << "OnWriteRemoteCharacteristicError() Error code: "
-                  << error_code.value();
+                  << static_cast<int>(error_code.value());
     service_ids_pending_eid_read_.erase(service_id);
     if (!has_callback_been_invoked_)
       NotifyFailureIfNoPendingEidCharReads();
